finally got limit with pagination working correctly with request.args.get

+++ blog/views.py
@@ -1,31 +1,32 @@
-from flask import render_template
+from flask import render_template, abort, redirect, request, url_for
 import mistune
 
 from . import app
 from .database import session, Entry
 
-@app.route("/")
-@app.route("/page/<int:page>")
-def entries(page=1, methods=["GET"]):
-    
-    try:
-        paginate_by = int(request.args.get('limit'))
-
-    except TypeError:
-        paginate_by = 5
+@app.route("/", methods=["GET","POST"])
+@app.route("/page/<int:page>", methods=["GET","POST"])
+@app.route("/?limit=<int:per_page>", methods=["GET","POST"])
+@app.route("/page/<int:page>?limit=<int:per_page>", methods=["GET","POST"])
+def entries(page=1):
         
     # Zero-indexed page
     page_index = page - 1
 
+    if request.args.get('limit'):
+        per_page = request.args.get('limit', type=int)  # if limit is added to the url
+    else:
+        per_page = 10  # default
+
     count = session.query(Entry).count()
     
-    if paginate_by > count:
-        paginate_by = count
+    if per_page > count:
+        per_page = count
 
-    start = page_index * paginate_by
-    end = start + paginate_by
+    start = page_index * per_page
+    end = start + per_page
 
-    total_pages = (count - 1) // paginate_by + 1
+    total_pages = (count - 1) // per_page + 1
     has_next = page_index < total_pages - 1
     has_prev = page_index > 0
 
@@ -38,7 +39,8 @@ def entries(page=1, methods=["GET"]):
         has_next=has_next,
         has_prev=has_prev,
         page=page,
-        total_pages=total_pages
+        total_pages=total_pages,
+        per_page=per_page
     )
     
 @app.route("/entry/add", methods=["GET"])
@@ -58,7 +60,7 @@ def add_entry_post():
     return redirect(url_for("entries"))
     
 @app.route("/entry/<int:entry_id>")
-def view_entry(entry_id):
+def single_entry(entry_id):
 	entry = session.query(Entry).filter(Entry.id==entry_id).first()
 	return render_template("single_entry.html",entry=entry)
 	
@@ -69,17 +71,30 @@ def edit_entry_get(entry_id):
 
 @app.route("/entry/<int:entry_id>/edit", methods=["POST"])
 def edit_entry_entry(entry_id):
+    if  request.method == "POST":
+        entry = session.query(Entry).get(entry_id)
+        entry.title=request.form["title"],
+        entry.content=mistune.markdown(request.form["content"])
+        session.add(entry)
+        session.commit()
+        return redirect(url_for("entries"))
+        
     entry = session.query(Entry).get(entry_id)
-    entry.title=request.form["title"],
-    entry.content=mistune.markdown(request.form["content"])
-    session.add(entry)
-    session.commit()
-    return redirect(url_for("entries"))
-  
+    return render_template("edit_entry.html", entry=entry)
+    
 @app.route("/entry/<int:entry_id>/delete", methods=["GET","POST"])
-def delete_entry(entry_id):
-    entry = session.query(Entry).filter(Entry.id==entry_id).first()
+def delete_entry_get(entry_id=1):
+    entry = session.query(Entry).get(entry_id)
+    if entry is None:
+        abort(404)
+    return render_template("delete_entry.html", entry=entry)
+
+@app.route("/delete/confirmation/<int:entry_id>", methods=["GET"])
+def delete_entry_confirmed(entry_id):
+    entry = session.query(Entry).get(entry_id)
+    if entry is None:
+        abort(404)
     session.delete(entry)
     session.commit()
-    return redirect(url_for("entries"))
+    return render_template("delete_confirmation.html")
